Numeric type audit: network, consensus, provider, rpc-types (#454)
* num_type_audit: network, consensus, provider, rpc-types

* `gas_limit` -> `u128`
* `gas` -> `u128`
* `gas_price` -> `u128`
* `max_fee_per_blob_gas` -> `u128`
* `transaction_type` -> `u8`
* `max_fee_per_gas` -> `u128`
* `max_priority_fee_per_gas` -> `u128`
* Consequently, changed fn args and return values accordingly.
* Satisfies clippy, tests will fail.

* fix(rpc-type): transaction requests tests

* fix(rpc-types): tests - transaction type

* add u128_hex_or_decimal_opt serde

* fix(providers): provider::tests

* fix(providers): layers::gas::tests

* fix: docs ci

* fix(provider): estimate_gas should return u128

* add(serde): unit test for u128_hex_or_decimal_opt

* fix(rpc-types, provider): use primitives in FeeHistory and its related methods

* add u128_vec serde methods

* ci nits

* nit

+++ crates/consensus/src/transaction/envelope.rs
@@ -484,7 +484,7 @@ mod tests {
             chain_id: u64::MAX,
             nonce: u64::MAX,
             gas_price: u128::MAX,
-            gas_limit: u64::MAX,
+            gas_limit: u128::MAX,
             to: TxKind::Call(Address::random()),
             value: U256::MAX,
             input: Bytes::new(),
